requirements.txt
memphis-py==1.3.1
python-decouple==3.8
backoff==2.2.1
loguru==0.7.2